requirements.txt
torchvision
git+https://github.com/facebookresearch/segment-anything.git@main
fastapi
pydantic
opencv-python
uvicorn
requests
Pillow
numpy
python-multipart